+++ api/app/Http/Controllers/TimetableController.php
@@ -1,53 +1,52 @@
 <?php
+
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Models\Timetable;
+use Illuminate\Http\Request;
 
-class TimetableController extends Controller {
-    public function index() {
-        $timetables = Timetable::withTrashed()->get();
-        return view('timetables.index', compact('timetables'));
+class TimetableController extends Controller
+{
+    public function __construct()
+    {
+        $this->authorizeResource(Timetable::class, 'timetable');
     }
 
-    public function create() {
-        return view('timetables.create');
+    public function index()
+    {
+        return Timetable::all();
     }
 
-    public function store(Request $request) {
-        // Lógica para criar um novo horário
-        $data = $request->all();
-        Timetable::create($data);
-        return redirect()->route('timetables.index');
+    public function showDay()
+    {
+        return response()->json(['day' => date('l')]);
     }
 
-    public function show($id) {
-        $timetable = Timetable::withTrashed()->find($id);
-        return view('timetables.show', compact('timetable'));
+    public function store(Request $request)
+    {
+        $this->authorize('create', Timetable::class);
+        $timetable = Timetable::create($request->all());
+        return response()->json($timetable, 201);
     }
 
-    public function edit($id) {
-        $timetable = Timetable::find($id);
-        return view('timetables.edit', compact('timetable'));
+    public function show(Timetable $timetable)
+    {
+        return $timetable;
     }
 
-    public function update(Request $request, $id) {
-        // Lógica para atualizar um horário
-        $data = $request->all();
-        $timetable = Timetable::find($id);
-        $timetable->update($data);
-        return redirect()->route('timetables.index');
+    public function update(Request $request, Timetable $timetable)
+    {
+        $this->authorize('update', $timetable);
+
+        $timetable->update($request->all());
+        return response()->json($timetable, 200);
     }
 
-    public function destroy($id) {
-        $timetable = Timetable::find($id);
-        $timetable->delete(); // Soft Delete
-        return redirect()->route('timetables.index');
-    }
+    public function destroy(Timetable $timetable)
+    {
+        $this->authorize('delete', $timetable);
 
-    public function restore($id) {
-        $timetable = Timetable::withTrashed()->find($id);
-        $timetable->restore(); // Restaurar um horário excluído
-        return redirect()->route('timetables.index');
+        $timetable->delete();
+        return response()->json(null, 204);
     }
 }
